LPS-196226 Change JSBundleConfigRegistry to an Util class, rename it to JSBundleConfigRegistryUtil and apply to usages
